Fix float conversion warning in String::toFloat.

+++ cores/esp32/WString.cpp
@@ -910,7 +910,7 @@ long String::toInt(void) const {
 
 float String::toFloat(void) const {
   if (buffer()) {
-    return atof(buffer());
+    return static_cast<float>(atof(buffer()));
   }
   return 0;
 }
